Remove symbols for last 3 libraries that have them: gl, al, html5
Rather than including these libraries on the link line based on running
`llvm-nm` on the inputs we now include them unconditionally.  This
behavior can be disabled using AUTO_NATIVE_LIBRARIES=0.

This change should not any effect on normal build since including these
libraries should be a nps.  For MAIN_MODULE=1 build it could be a
slightly increased code size, but MAIN_MODULE=1 builds are already
enormous and MAIN_MODULE=2 is the path to make them smaller.

Since these are the last libraries to use these symbols files we can
also remove all the code surrounding the processing of them!